requirements.txt
aiogram==3.0.0rc1
aioredis==2.0.1
alembic==1.11.3
asyncpg==0.28.0
SQLAlchemy==2.0.19
redis==4.6.0
aiohttp~=3.8.3
PyYAML~=6.0.1
cachetools~=5.3.1
colorlog==6.7.0